refactor: Rename `layouts` -> `views`
